fix: added authentication_required to authentication in common [DET-3316]
