+++ cli/determined_cli/master.py
@@ -6,7 +6,7 @@ from typing import Any, List
 from requests import Response
 
 from determined_common import api
-from determined_common.api.experiment import authentication_required
+from determined_common.api.authentication import authentication_required
 from determined_common.check import check_gt
 
 from .declarative_argparse import Arg, Cmd
