+++ cli/determined_cli/template.py
@@ -7,7 +7,7 @@ from ruamel import yaml
 from termcolor import colored
 
 from determined_common import api
-from determined_common.api.experiment import authentication_required
+from determined_common.api.authentication import authentication_required
 
 from . import render
 from .declarative_argparse import Arg, Cmd
