+++ cli/determined_cli/trial.py
@@ -5,7 +5,7 @@ from typing import Any, List
 
 from determined_cli import render
 from determined_common import api
-from determined_common.api.experiment import authentication_required
+from determined_common.api.authentication import authentication_required
 from determined_common.experimental import Determined
 
 from .checkpoint import format_checkpoint, format_validation, render_checkpoint
